@@ -34,9 +34,9 @@
     "@types/carbon-components-react": "^7.10.9",
     "@types/jest": "^25.2.3",
     "@types/lodash-es": "^4.17.3",
-    "@types/react": "^16.14.5",
-    "@types/react-dom": "^16.9.12",
-    "@types/react-router-dom": "^5.1.7",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-router-dom": "^5.3.3",
     "@types/uuid": "^8.3.0",
     "@types/webpack-env": "^1.16.0",
     "@types/yup": "^0.29.11",
@@ -60,8 +60,8 @@
     "openmrs": "next",
     "prettier": "^2.4.1",
     "pretty-quick": "^2.0.2",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-i18next": "^11.0.0",
     "react-router-dom": "^5.0.0",
     "rxjs": "^6.6.3",
