@@ -44,7 +44,7 @@
     "@openmrs/esm-framework": "3.x",
     "carbon-components": "10.x",
     "carbon-icons": "7.x",
-    "react": "16.x",
+    "react": "^18.1.0",
     "react-i18next": "11.x",
     "react-router-dom": "5.x",
     "rxjs": "6.x"
