@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import Root from './root.component';
 import { match } from 'react-router-dom';
 
@@ -16,7 +16,8 @@ const sampleMatchProp: match<{ patientUuid: string }> = {
 describe('root component', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(
+    const root = createRoot(div);
+    root.render(
       <Root
         savePatientForm={jest.fn()}
         match={sampleMatchProp}
@@ -25,7 +26,6 @@ describe('root component', () => {
         patientIdentifiers={[]}
         relationshipTypes={{ results: [] }}
       />,
-      div,
     );
   });
 });
